chore: restructured metal file/folder structure
