+++ metal/terrain/Biome.h
@@ -8,13 +8,6 @@
 #ifndef Biome_h
 #define Biome_h
 
-#ifdef __METAL_VERSION__
-#define NS_ENUM(_type, _name) enum _name : _type _name; enum _name : _type
-#define NSInteger metal::int32_t
-#else
-#import <Foundation/Foundation.h>
-#endif
-
 #include <simd/simd.h>
 
 /// Defines a biome, consisting of a elevation and color that defines a particular biome
@@ -36,30 +29,6 @@ typedef struct Biome {
     /// elevation 0.6 will range from 0.6-blendRange/2 to 0.6+blendRange/2).
     float blendRange;
     
-    /// Whether trees should appear on this biome
-    bool hasTrees;
-    
 } Biome;
-
-/// Represents a type of biome with associated color
-typedef NS_ENUM(NSInteger, BiomeType) {
-    BiomeTypeOcean = 1,
-    BiomeTypeShallowWater,
-    BiomeTypeShore,
-    BiomeTypeScorched,
-    BiomeTypeBare,
-    BiomeTypeTundra,
-    BiomeTypeSnow,
-    BiomeTypeTemperateDesert,
-    BiomeTypeShrubland,
-    BiomeTypeTaiga,
-    BiomeTypeGrassland,
-    BiomeTypeTemperateDeciduousForest,
-    BiomeTypeTemperateRainforest,
-    BiomeTypeSubtropicalDesert,
-    BiomeTypeTropicalSeasonalForest,
-    BiomeTypeTropicalRainforest,
-    BiomeTypeTotal,
-};
 
 #endif /* Biome_h */
